Adding cast to list and map size comparisons to avoid signed/unsigned comparison:
    BOOST_CHECK_EQUAL(list.size(), (size_t) 0);


git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@997846 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/tests/Address.cpp
@@ -92,7 +92,7 @@ QPID_AUTO_TEST_CASE(testParseOptionsWithEmptyList)
     Address address("my-topic; {a:[], x:101}");
     BOOST_CHECK_EQUAL(std::string("my-topic"), address.getName());
     Variant::List& list = address.getOptions()["a"].asList();
-    BOOST_CHECK_EQUAL(list.size(), 0);
+    BOOST_CHECK_EQUAL(list.size(), (size_t) 0);
     BOOST_CHECK_EQUAL((uint16_t) 101, address.getOptions()["x"].asInt64());
 }
 
@@ -101,7 +101,7 @@ QPID_AUTO_TEST_CASE(testParseOptionsWithEmptyMap)
     Address address("my-topic; {a:{}, x:101}");
     BOOST_CHECK_EQUAL(std::string("my-topic"), address.getName());
     Variant::Map& map = address.getOptions()["a"].asMap();
-    BOOST_CHECK_EQUAL(map.size(), 0);
+    BOOST_CHECK_EQUAL(map.size(), (size_t) 0);
     BOOST_CHECK_EQUAL((uint16_t) 101, address.getOptions()["x"].asInt64());
 }
